get users -> get consultants and error handling added
